added hasSignUp field to users in migration

+++ app/database/migrations/2015_03_03_172810_add_user_validation.php
@@ -14,7 +14,8 @@ class AddUserValidation extends Migration {
 	{
 		Schema::table('users', function(Blueprint $table)
 		{
-			$table->integer('verified');
+            $table->boolean('verified');
+			$table->boolean('hasSignUp');
 		});
 	}
 
@@ -28,7 +29,15 @@ class AddUserValidation extends Migration {
 	{
 		Schema::table('users', function(Blueprint $table)
 		{
-			$table->dropColumn('verified');
+            if (Schema::hasColumn('users', 'verified'))
+            {
+                $table->dropColumn('verified');
+            }
+            
+            if (Schema::hasColumn('users', 'hasSignUp'))
+            {
+                $table->dropColumn('hasSignUp');
+            }
 		});
 	}
